# frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blood Bank API",
    "description": "API for the Blood Bank project",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/bloodbank/{bloodbank_id}/appointment": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Bloodbank Appointments",
        "description": "Show appointments for blood bank.\n\nBy default only the ones scheduled after 00:00 in Oslo on the same day",
        "operationId": "bloodbank_appointments",
        "parameters": [
          {
            "name": "bloodbank_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Bloodbank Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "name": "show_cancelled",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Show Cancelled"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingSlotType"
                  },
                  "title": "Response Bloodbank Appointments Admin Bloodbank  Bloodbank Id  Appointment Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/appointment/{appointment_id}/note": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Appointment Add Note",
        "operationId": "appointment_add_note",
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNoteRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/appointment/{appointment_id}/result": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Testresult For Appointment Admin",
        "operationId": "get_testresult_for_appointment_admin",
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Donation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/appointment/{appointment_id}": {
      "patch": {
        "summary": "Update",
        "description": "Update an appointment, either changing its booking slot or cancelling it altogether\n\nThis action can be performed by the donor or an admin at the blood bank(s) in question\n\nTODO:\n- validate that requested bookingslot works\n- add appointment note entry documenting the change",
        "operationId": "update",
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAppointmentRow"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/appointment/{appointment_id}/note": {
      "post": {
        "summary": "Add Note",
        "operationId": "add_note",
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNoteRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/testresult/{testresult_id}": {
      "get": {
        "summary": "Get Testresult",
        "operationId": "get_testresult",
        "parameters": [
          {
            "name": "testresult_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Testresult Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/InterviewDetailsRow"
                    },
                    {
                      "$ref": "#/components/schemas/DonationTestDetailsRow"
                    },
                    {
                      "$ref": "#/components/schemas/EntryFormDetailsRow"
                    }
                  ],
                  "title": "Response Get Testresult Testresult  Testresult Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/testresult/appointment/{appointment_id}": {
      "get": {
        "summary": "Get Testresult For Appointment",
        "operationId": "get_testresult_for_appointment",
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Donation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddNoteRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "AddNoteRequest"
      },
...
      "AppointmentTypeAdmin": {
        "properties": {
          "appointment_id": {
            "type": "integer",
            "title": "Appointment Id"
          },
          "appointment_cancelled": {
            "type": "boolean",
            "title": "Appointment Cancelled"
          },
          "donor_id": {
            "type": "integer",
            "title": "Donor Id"
          },
          "donor_blood_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BloodType"
              },
              {
                "type": "null"
              }
            ]
          },
          "donor_name": {
            "type": "string",
            "title": "Donor Name"
          },
          "donor_email": {
            "type": "string",
            "title": "Donor Email"
          },
          "donor_phone": {
            "type": "string",
            "title": "Donor Phone"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/NoteType"
            },
            "type": "array",
            "title": "Notes"
          },
          "donations": {
            "items": {
              "$ref": "#/components/schemas/DonationType"
            },
            "type": "array",
            "title": "Donations"
          }
        },
        "type": "object",
        "required": [
          "appointment_id",
          "appointment_cancelled",
          "donor_id",
          "donor_blood_type",
          "donor_name",
          "donor_email",
          "donor_phone",
          "notes",
          "donations"
        ],
        "title": "AppointmentTypeAdmin"
      },
      "AppointmentUpdateRequest": {
        "properties": {
          "bookingslot_id": {
            "type": "integer",
            "title": "Bookingslot Id"
          },
          "cancelled": {
            "type": "boolean",
            "title": "Cancelled"
          }
        },
        "type": "object",
        "required": [
          "bookingslot_id",
          "cancelled"
        ],
        "title": "AppointmentUpdateRequest"
      },
...
      "BloodType": {
        "type": "string",
        "enum": [
          "O+",
          "O-",
          "A+",
          "A-",
          "B+",
          "B-",
          "AB+",
          "AB-"
        ],
        "title": "BloodType"
      },
...
      "BookingSlotType": {
        "properties": {
          "bookingslot_id": {
            "type": "integer",
            "title": "Bookingslot Id"
          },
          "bookingslot_time": {
            "type": "string",
            "format": "date-time",
            "title": "Bookingslot Time"
          },
          "bookingslot_duration": {
            "type": "string",
            "format": "duration",
            "title": "Bookingslot Duration"
          },
          "bookingslot_remaining_capacity": {
            "type": "integer",
            "title": "Bookingslot Remaining Capacity"
          },
          "appointments": {
            "items": {
              "$ref": "#/components/schemas/AppointmentTypeAdmin"
            },
            "type": "array",
            "title": "Appointments"
          }
        },
        "type": "object",
        "required": [
          "bookingslot_id",
          "bookingslot_time",
          "bookingslot_duration",
          "bookingslot_remaining_capacity",
          "appointments"
        ],
        "title": "BookingSlotType"
      },
...
      "Donation": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "appointment_id": {
            "type": "integer",
            "title": "Appointment Id"
          },
          "amount_ml": {
            "type": "number",
            "title": "Amount Ml"
          },
          "is_blood_not_plasma": {
            "type": "boolean",
            "title": "Is Blood Not Plasma"
          }
        },
        "type": "object",
        "required": [
          "id",
          "appointment_id",
          "amount_ml",
          "is_blood_not_plasma"
        ],
        "title": "Donation"
      },
      "DonationTestDetailsRow": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "donor_id": {
            "type": "integer",
            "title": "Donor Id"
          },
          "form_id": {
            "type": "integer",
            "title": "Form Id"
          },
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time"
          },
          "validity_duration": {
            "type": "string",
            "format": "duration",
            "title": "Validity Duration"
          },
          "invalidated": {
            "type": "boolean",
            "title": "Invalidated"
          },
          "ok_to_donate": {
            "type": "boolean",
            "title": "Ok To Donate"
          },
          "donation_test_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Donation Test Id"
          },
          "donation_id": {
            "type": "integer",
            "title": "Donation Id"
          },
          "tester_admin_id": {
            "type": "integer",
            "title": "Tester Admin Id"
          },
          "appointment_id": {
            "type": "integer",
            "title": "Appointment Id"
          },
          "amount_ml": {
            "type": "number",
            "title": "Amount Ml"
          },
          "is_blood_not_plasma": {
            "type": "boolean",
            "title": "Is Blood Not Plasma"
          },
          "tester_admin_name": {
            "type": "string",
            "title": "Tester Admin Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "donor_id",
          "form_id",
          "time",
          "validity_duration",
          "invalidated",
          "ok_to_donate",
          "donation_test_id",
          "donation_id",
          "tester_admin_id",
          "appointment_id",
          "amount_ml",
          "is_blood_not_plasma",
          "tester_admin_name"
        ],
        "title": "DonationTestDetailsRow"
      },
...
      "DonationType": {
        "properties": {
          "donation_id": {
            "type": "integer",
            "title": "Donation Id"
          },
          "amount_ml": {
            "type": "number",
            "title": "Amount Ml"
          },
          "is_blood_not_plasma": {
            "type": "boolean",
            "title": "Is Blood Not Plasma"
          }
        },
        "type": "object",
        "required": [
          "donation_id",
          "amount_ml",
          "is_blood_not_plasma"
        ],
        "title": "DonationType"
      },
...
      "EntryFormDetailsRow": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "donor_id": {
            "type": "integer",
            "title": "Donor Id"
          },
          "form_id": {
            "type": "integer",
            "title": "Form Id"
          },
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time"
          },
          "validity_duration": {
            "type": "string",
            "format": "duration",
            "title": "Validity Duration"
          },
          "invalidated": {
            "type": "boolean",
            "title": "Invalidated"
          },
          "ok_to_donate": {
            "type": "boolean",
            "title": "Ok To Donate"
          },
          "entry_form_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Form Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "donor_id",
          "form_id",
          "time",
          "validity_duration",
          "invalidated",
          "ok_to_donate",
          "entry_form_id"
        ],
        "title": "EntryFormDetailsRow"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InterviewDetailsRow": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "donor_id": {
            "type": "integer",
            "title": "Donor Id"
          },
          "form_id": {
            "type": "integer",
            "title": "Form Id"
          },
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time"
          },
          "validity_duration": {
            "type": "string",
            "format": "duration",
            "title": "Validity Duration"
          },
          "invalidated": {
            "type": "boolean",
            "title": "Invalidated"
          },
          "ok_to_donate": {
            "type": "boolean",
            "title": "Ok To Donate"
          },
          "interview_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Id"
          },
          "interviewer_admin_id": {
            "type": "integer",
            "title": "Interviewer Admin Id"
          },
          "interviewer_admin_name": {
            "type": "string",
            "title": "Interviewer Admin Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "donor_id",
          "form_id",
          "time",
          "validity_duration",
          "invalidated",
          "ok_to_donate",
          "interview_id",
          "interviewer_admin_id",
          "interviewer_admin_name"
        ],
        "title": "InterviewDetailsRow"
      },
...
      "NoteType": {
        "properties": {
          "author_user_id": {
            "type": "integer",
            "title": "Author User Id"
          },
          "author_name": {
            "type": "string",
            "title": "Author Name"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time"
          }
        },
        "type": "object",
        "required": [
          "author_user_id",
          "author_name",
          "message",
          "time"
        ],
        "title": "NoteType"
      },
...
      "UpdateAppointmentRow": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "bookingslot_id": {
            "type": "integer",
            "title": "Bookingslot Id"
          },
          "cancelled": {
            "type": "boolean",
            "title": "Cancelled"
          },
          "donor_id": {
            "type": "integer",
            "title": "Donor Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "bookingslot_id",
          "cancelled",
          "donor_id"
        ],
        "title": "UpdateAppointmentRow"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}